fast/requirements.txt
# FastAPI RAG Backend Dependencies

# Web Framework
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6

# LangChain Core
langchain>=0.3.0
langchain-community>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.1.0

# Vector Store
chromadb>=0.4.22

# PDF Processing
pypdf>=4.0.0

# Environment
python-dotenv>=1.0.0

# Note: Create a .env file with your OpenAI API key
# OPENAI_API_KEY=your_api_key_here
